pyproject.toml
[project]
name = "anki-telegram"
version = "0.1.0"
description = "Telegram bot to create Anki flashcards via Copilot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "github-copilot-sdk",
  "python-telegram-bot>=21.0",
  "PyYAML>=6.0",
  "httpx>=0.27.0",
  "pydantic>=2.6.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.5.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
